requirements.txt
accelerate
bitsandbytes
datasets
deepspeed==0.15.0
einops
ipykernel
isort
jaxtyping
jsonlines
loralib
optimum
packaging
peft
plotly
ray[default]
tensorboard
torchmetrics
tqdm
transformers
transformers_stream_generator
transformer_lens
wandb
wheel